requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-heroku==0.3.1
django-render-block==0.10
django-templated-email==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
